+++ Assets/_Scripts/GameManager.cs
@@ -19,5 +19,11 @@ namespace _Scripts
             }
 
         }
+
+        private void Start()
+        {
+            Cursor.lockState = CursorLockMode.Locked;
+            Cursor.visible = false;
+        }
     }
 }